Added new script v2-samplesheet-to-json
Writes a v2 samplesheet to a json object.

This is particularly useful if a samplesheet needs to be quickly modified.

Other updates

Ensure analysis urns already in cloud-settings aren't overwritten

Fixed barcode_mismatches_index1 to barcode_mismatches_index_1,to be consistent with adapter_read_1, same applies for read_1_cycles etc.

Only time that index<num> is for index1 and index2 with no prefix or suffix